Use the new unordered maps and sets when using castxml
Fixes the following warnings:
warning Use of the header <ext/hash_map> is deprecated. Migrate to <unordered_map>
warning Use of the header <ext/hash_set> is deprecated. Migrate to <unordered_map>

The old includes are kept for backward compatibility, especially the Travis builds which still use gccxml for the moment.