Remove xml and c++ files and the gccxml-0.9-upgrade folder
These files were added to illustrate the changes between
gccxml 0.7 and gccxml 0.9. Support for GCCXML will be dropped
in one of the next versions and there is no reason to keep
the files around in the tree. They were not used in any
tests. Besides; they are still present in the git tree
if needed.