Move the joining of declarations from the project reader to the source reader.
When using pygccxml with CastXML, the following test failed:

======================================================================
FAIL: test_correctness (__main__.tester_t)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "unittests/project_reader_correctness_tester.py", line 61, in test_correctness
    self.__test_correctness_impl(src)
  File "unittests/project_reader_correctness_tester.py", line 57, in __test_correctness_impl
    file_name)
AssertionError: There is a difference between declarations in file core_types.hpp.

This happened because "#include <iostream>" brings some methods which are defined multiple
times (__test()). These are probably internal methods which we don't care about; but the Clang
exposes them in the AST and they are present in the xml files.
The difference in the resulting printed tree in pygccxml (between source and project reader)
was due to the fact that the source reader did not join these declarations.

This patches does the following thing:
- Move the declaration joining code to the source reader
- Join the declarations in the source reader
- When using the project reader; skip the joining of the declarations in the source reader,
by providing the join_decls=False argument to the source reader. This is important for
performance as the joining should happen only once on all declarations.