Add option to remove the __va_list_tag declaration from the tree when parsing with CastXML
This option is on by default.

The __va_list_tag declarations are internal declarations, which are probably
not needed. They are for example polluting the declarations tree when running
pyplusplus. The removal can still be re-enabled by setting the utils.remove__va_list_tag
option to False.

This fixes also the filter test, which has now the correct initial number of declarations
(the same as for gccxml). It also removes the hack in the hierarchy_travelling test.