Fix for typedef'd unnamed class/struct
Unlike GCCXML, CastXML creates two distinct XML nodes for a typedef of
the form

            typedef (class|struct) {} foo;

The first XML node is an unnamed class/struct and the second is a
typedef that refers to the first node. This requires special handling
during parsing so that the name of the typedef is propagated to the
class/struct.

Change-Id: I91cc3753d50d41ff92c945a7b9801a0108163e6a