Fix unqualified integral const in default arguments for CastXML
CastXML, prior to XML output versin 1.137, didn't modify integral
constant names used in default arguments for functions, it simply used
the same text present in the source file. This causes compilation
failures in bindings generated by pyplusplus in case of code such as the
following:

	struct st
	{
		static int const DEFAULT = 0;
		void fun(int arg=DEFAULT);
	};

In this case, CastXML produced `DEFAULT` as the default value while
GCCXML qualifies the name and sets the default value to `st::DEFAULT`.

Consequently, pyplusplus refers to the default value as `DEFAULT`, which
fails compilation because the name needs to be fully qualified as
`st::DEFAULT`. This fix produces the desired qualified output from
pygccxml with both GCCXML and CastXML.

As of XML output version 1.137, CastXML now produces fully qualified
names for default arguments.

Change-Id: Ie22ae5e7e058743efcd986a44797005be8e9c891