Adapt patched output to CastXML cvs_revision 1.137
CastXML's output for default arguments for function parameters was
modified in response to CastXML/CastXML#51. The tool now generates fully
qualified names in default arguments, so we need to check the
`cvs_revision` attribute in the XML output to differentiate between the
different output formats.

Change-Id: I18ee4d2319dd879e8b9a2d40bf6bee2d3ec1224e

Cherry-picked from the develop branch