Fix relinking of unnamed classes (issue #45)
With gcc5 + castxml + std=c++11 some code was not
parseable by pygccxml (see previous commit).

This fixes the problem, but is probably not the
final solution, but as the other cases in the relink
method, a better understanding of this code part
should hopefully help find a better strategy to
handle these cases.