Remove unnecessary try/catch block
The exception clause is too broad, as it is catching all exceptions.

The git history does not help to understand why that try/catch block is there.

If there is an exception, we want to know what it is and act accordingly.
I do not expect any exception in that code block (even if an IOError is still
possible, but this could happen at a lot of places in the code).