Fix patching of enums in default arguments for C++03
Given the following code

    namespace ns
    {
        enum color {red, green, blue};
        void test( color arg=blue );
    }

CastXML produces a fully qualified name for the default argument -
`::ns::color::blue`.

While this default argument string is valid for C++11 and later, and
indeed required if `color` were a scoped enumeration, the default
argument is invalid for C++03 code.

Prior to C++11, enumerator names were added to the enclosing scope, so
they could not be qualified using the enum-name. Hence, for C++03, the
default argument needs to be patched to say `::ns::blue`.

Change-Id: I72c5d4fc53c5b5dc6f5fa38d10cb5484ba3bf85a