Fix failing tests under ubuntu wily (see #47)
The __castxml__ definition does not exist with the castxml shipped with wily.
In this case, do not use it and check for __GCCXML__ instead, to know which xml generator we are using.

The travis build was not getting this error because we are using one of the latest versions of CastXML there
instead of the one shipped by apt.

Reorganised the definitions to make more sense.