Remove unused __make_flatten_generator function
This is used nowhere and not tested. Also, this method is marked as "private" with __,
so it won't go through the deprecation process.

It is not used in py++ and I had a quick look on github, and found no application.
If somebody needs an equivalent functionality the function can be rewritten properly + tested later.