Move some functions from algorithm module to new declaration_utils module
In the algorithm module, some functions are strongly linked to scopedef classes.
Algorithm is imported at a lot of places, which created some potentials conflicts and
made refactoring difficult. Now, instead of importing algorithm everywhere,
only the module with the 4 needed functions is imported. This module has the
advantage that it has no other dependencies, making it self-contained.