Move the find_copy_constructor and find_trivial_constructor to the type_traits module
In fact, there was already a find_trivial_constructor function in the type_traits module,
so this makes the situation cleaner by providing only one way to do things.

Deprecate the find_copy_constructor and find_trivial_constructor methods from class_t

Do not print information about copy constructors for constructor_t (this is a side-effect of the changes).
That information would be nice to have, but not necessary.