Split the type_traits module in two
One part, which remained in the type_traits module, are the basic
functions needed to work with type_traits. These functions need very
few dependenies.

The second part are all the functions which are related to classes.
This allows now to import the type_traits module mostly everywhere
without cyclic import problems. The type_traits_classes is then the
more specialized version from the type_traits module.

The split makes also sense as the type_traits module was very long,
and the logical split is welcome.