Extract some more traits methods to a separate module
The auto_ptr_traits, smart_pointer_traits and internal_type_traits are only imported in the __init__.py file,
so they can easily be moved to a separate module (called other_traits).
This reduces the number of imports in container_traits.

The impl_details class is also moved to a separate module.