Imrove support for newer gccxml debian packages
For backward compatibility, the debian packages wrapped
CastXML inside the gccxml package. This means that if
you call gccxml on ubunut 16.04 for example, what you
get is CastXML ... this is a very weird decision.
I can't support that hybrid package, because it will mess up
pygccxml and a bunch of code would need to be added for this.

People are now exlicitely asked to tell what they want
to use: either they use the castxml package, which contains
the CastXML binary, or they use the gccxml.real binary,
which is provided by the gccxml package.