Performance: do not check for string in patter_parser (#56)
We can trust that the input is a string. Checking for this is expensive.
The number of calls to is_str went down from 280000 calls to 22000 in most simple cases.

The test is also no more needed. If somebody gives a wrong input, it will just fail with an exception.