Introduce new class to store byte information
The byte_align and byte_size getters and setters were duplicated.
Using a common byte_info class to store that information helps keeping the code at one place.

One disadvantage is that the type_t is now a child class of byte_info, which is not strictly
speaking part of the traditional objects we have in pygccxml.

But this does not hinder the type checks, and the first type that is printed when printing such objects
is the one from one of the subclasses, so that the byte_info class should not be visible to the users,
only if they print explicitly the MRO.