Raise NotImplementedError() instead of raise NotImplemented()
Maybe there was a confusion with the NotImplemented builtin, which probably makes sense in
a comparison scheme between objects (we are in the _get__cmp__items method here) (often used for performance reasons).

Anyway, this method is correctly overridden by the current 6 child classes, so NotImplementedError() was probably ment here initially.