Do not reset xml_generator to None during the test
The order of the tests has changed as I moved some tests to the end.
Setting this value to None breaks the following tests.

The goal for version 1.9.0 will be to go back to a xml generator saved
in the config, instead of this global variable which is a pain to manage