Add coverage report for windows
An “empty” build is added to the build matrix, which sets
the WINDOWS env variable to 1.

When WINDOWS=1, the test suite is not run, and the coverage
results are fetched from the appveyor build using the appveyor-artifacts tool.
This allows to create a “fake” windows build on travis, which serves
as a simple proxy for the coverage report.