Move declarations joining code to separate file
This code does not really belong to the source_reader_t class.

The source_reader_t is used to read the sources.
Joining/cleaning up the declarations is a different topic.

The join_declarations method was made visible
through previous refactorings, which was probably
wrong. A deprecation warning has been added
just in case, so that users can report if they really
need this method (which I don’t believe will happen).