Use new syntax for super hash calls
This fixes the following error with pypy 5.6.0 (2.7.12)

File "../pygccxml/pygccxml/declarations/calldef.py", line 233, in __hash__
    hash(self.return_type) ^
TypeError: unbound method __hash__() must be called with super instance as first argument (got free_function_t instance instead)

The cpython implementation was OK with this, though the syntax is dubious.
The new way of writing this is better as it clearly defines what is called
in the class tree.

Some of the calls may even be unnecessary, but I am leaving
them as-is for the moment.

Cherry-picked from the develop branch