Regroup runtime exceptions
First, the two exceptions contained in the matcher class are extracted to a new file.
The exception in the algorithm.py file joins both exceptions in this new file.

The exceptions were already in the public API, so we keep this as is.

In a second step the two exceptions which were re-declared in
the scopedef_t will need to be deprecated. There is no reason
that this class exposes both exceptions if they are exposed in
the pygccxml.declarations package.