Do not redefine variable type. Simplify logic
Fixes:
Redefinition of found type from pygccxml.declarations.cpptypes.declarated_t to pygccxml.declarations.cpptypes.const_t (redefined-variable-type)

Also, return directly when a declaration is found, instead of keeping a found variable around.
Note that this part of code could be improved, the method is way to big and the logic too convoluted.