Remove superfluous logic in __remove_alias method
The remove_alias method already makes a check for typedef.typedef_t
and gets the decl_type. Then it calls the "private" __remove_alias
method. This one does not need a supplementary check.

__remove_alias can call itself recursively, but the argument would
either be a new decl_type which is not a typedef_t (first case), or
the .base of a compount_t, which is also never a typedef_t.

The coverage also tells us that this condition is never called,
and we have extensive checks for that part of the code.
This makes me confident that this condition can be removed.

Obviously the gain is to have a bunch of insistence checks less,
as remove_alias can be called thousands of times.