Completely remove compiler attribute from declarations
The attribute (and the getters/setters) were removed in 1.8.0.
I forgot to remove the test, and as the attribute was still set
from outside in the scanner part, the test was still good.

Lessons learned: setting an attribute like this is probably
not a good idea, because I lost of time looking for this
compiler attribute which was already removed ...