The configuration now holds a xml_generator_from_xml_file object
This will allow to modify the configuration objects passed to
the parser methods in __init__py and store the xml_generator_from_xml_file object.
Ideally I would like to return xml_generator_from_xml_file directly, but this would
break the API because these methods only return the declarations list.
It is a sort of "hack" of the current infrastructure to still be able to read the
xml_generator_from_xml_file object from the place where the call methods get called.