Use the xml_generator_from_xml_file object in the core_tester
This demonstrates the usage of the xml_generator_from_xml_file object.
The ultimate goal to stop using global variables in utils has been reached.
All the other calls in the others tests will need to be updated too, this
change here is just to demonstrate the proof of concept.