Do not store logger as attribute
During the last refactorings I made that change,
but it is not a good one, as the object gets pickled,
and this will fail in some situations (especially
under windows). Just passing the logger through is enough anyway.

Also, having it as an attribute exposes it in the public API,
which is probably not what I wanted to do.