Add elaborated type prefix to declaration strings
For this a elaborated_info class is added to avoid code duplication
between the class_t and enumeration_t classes.

The declaration string building of the elaborated_t class just checks if this
attribute is defined and uses it directly. This is more elegant/performant than
adding isinstance checks here to check for class_t and enumeration_t.

I am not really happy with the ordering of the declaration string,
as it is for example: "declname class const", but this seems to
always have been like this. I will investigate this later.