Revert "Put qualifiers on the left side of declarations string"
This reverts commit 16c536031407f25ee0b712177197025c31e94269.

Moving the qualifiers to the left side was wrong.
Having them on the left side is sometimes OK, but not in
all the cases. As we want to keep the decl_string methods
as simple as possible, pygccxml will keep the convention
that qualifiers will be on the right side of the base type.