Make apply_visitor ready for the class names refactoring
The classes will be written in CamelCase, and the _t will
be dropped. For this, the visitor mechanism needs to
be adapted. The regex is probably slower but compatible
with both the old and new syntax. This will stay like
this until the refactoring is done. The temporary solution
can then be dropped.