Add __hash__ delegation back
In commit 0ece970e03dac49e996466bb841daabdbce986a3, useless
redefintions of the __hash__ methods were removed, as they
were flagged as such by pylint.

In fact, all these classes defined the __eq__ method.
In that case, the __hash__ methods would have returned None.

As external tools (like pyplusplus) rely on the fact that the __hash__
methods return something sensible for all types and declaration objects,
we need to defined these methods.

The test_hash test class has been added to make sure that
the __hash__ methods are properly implemented, as this is
part of the public API of pygccxml.

The type_qualifiers_t class will get it's __hash__ method implemented
in a separate commit, as it needs a separate test.