Add caching of file with new dictionary
Add __files_text to the scanner object.  It is a dictionary of file
declaration to the content of each file in a list of strings.

The first time a comment is a part of a file, read it in and store it in
the dictionary.  Each time afterwards, use the stored value to capture
the lines.

Use the (begin/end) columns to more precisely capture the text of each
comment.

Update the test to remove extra characters.