Add standard error to exception message
The rationale is that when castxml has an error it raises a RuntimeError to the user with the message.
However, the message only contains the stdout from castxml - this makes the RuntimeError contain no information about what actually went wrong.
This change allows the user to add error handling to their applications based on what went wrong with castxml (such as changing their include directories, etc)